Integrated SCORM-Export in User Interface
This overrules xowikis standard admin page (index).
The only thing added is the SCORM-Export button, which
redirects to the url admin/export-cp.tcl (the actual
export script)

Also, we override the resolve_page method of the standard
xowiki package, because it only delivers files via
download/file urls. but we need the files also delivered
via normal xowikiinstance/abc.jpg URLs, because imported
content packages of course link to their assets directly.
thus, we also want to search for xowiki-files (not only pages)
when an request to the "root-folder" arrives